+++ src/main/java/com/sg/soft/KataBankApp/service/StatementService.java
@@ -2,12 +2,13 @@ package com.sg.soft.KataBankApp.service;
 
 import com.sg.soft.KataBankApp.dto.StatementDTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface StatementService {
     List<StatementDTO> findAllByAccountNumber(Long accountNumber, Integer page, Integer size);
 
-    StatementDTO makeDeposit(Long accountNumber, Double amount);
+    StatementDTO makeDeposit(Long accountNumber, BigDecimal amount);
 
-    StatementDTO makeWithdrawal(Long accountNumber, Double amount);
+    StatementDTO makeWithdrawal(Long accountNumber, BigDecimal amount);
 }
